extend test base; add more defaults to config to be used in project_manager
