+++ .travis.yml
@@ -4,6 +4,13 @@ python:
   - "3.6"
 before_install:
   - sudo apt-get install -y bedtools
+  - sudo sh -c 'echo "deb http://cran.rstudio.com/bin/linux/ubuntu trusty-cran35/" >> /etc/apt/sources.list'
+  - gpg --keyserver keyserver.ubuntu.com --recv-key E084DAB9
+  - gpg -a --export E084DAB9 | sudo apt-key add -
+  - sudo apt-get update
+  - sudo apt-get -y install r-base
+  - sudo su - -c "R -e \"if (!requireNamespace('BiocManager', quietly = TRUE)){install.packages('BiocManager', repos='https://CRAN.R-project.org')}; BiocManager::install('DESeq2', version = '3.8')\""
+
 install:
   - pip install -r requirements/requirements.txt
   - pip install .
